fixes the working dirs for all scripts

+++ scripts/install.sh
@@ -1,30 +1,52 @@
 #!/bin/bash
 set -euo errtrace
 
-# Handle errors.
+# Handle errors
 error_handle() {
     echo "error: ${BASH_SOURCE[1]} at line ${BASH_LINENO[0]}"
     echo "An unexpected error has occured."
 }
 trap error_handle ERR
 
-# Check if aftman is installed.
-if !command -v aftman >/dev/null 2>&1; then
-    # Ask the user if they want to install aftman.
+# Ensure we are in the correct directory
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)"
+cd $SCRIPT_DIR
+
+# Options
+GITKEEP=false
+
+for i in "$@"; do
+  case $i in
+    -gk|--gitkeep|--git-keep)
+      GITKEEP=true
+      shift
+      ;;
+    -*|--*)
+      echo "Unknown Option $i"
+      exit 1
+      ;;
+    *)
+      ;;
+  esac
+done
+
+# Check if aftman is installed
+if ! command -v aftman >/dev/null 2>&1; then
+    # Ask the user if they want to install aftman
     read -p "Aftman is not installed. Do you want to install it automatically? (Y/N) " answer
     case $answer in
         y|Y )
-            # Check the user's operating system.
+            # Check the user's operating system
             OS=$(uname | tr '[:upper:]' '[:lower:]')
             case $OS in
                 msys* | mingw* | cygwin* ) # Windows (Cygwin/MSYS)
-                    # Download the executable.
+                    # Download the executable
                     curl -L -o aftman.zip "https://github.com/LPGhatguy/aftman/releases/download/v0.2.8/aftman-0.2.8-windows-x86_64.zip"
                     # Extract the file.
                     unzip aftman.zip
-                    # Run the installation command.
+                    # Run the installation command
                     ./aftman self-install
-                    # Remove the downloaded files.
+                    # Remove the downloaded files
                     rm aftman.exe
                     rm aftman.zip
                     ;;
@@ -45,18 +67,20 @@ if !command -v aftman >/dev/null 2>&1; then
     esac
 fi
 
-# Remove .git tracking.
-rm -r .git
+# Remove .git tracking
+if ! $GITKEEP; then
+  rm -r .git
+fi
 
-# Install dependencies.
+# Install dependencies
 aftman install --no-trust-check
 wally install
 git submodule update --init
 
-# Build the project.
+# Build the project
 rojo sourcemap default.project.json -o sourcemap.json
 wally-package-types --sourcemap sourcemap.json Packages
 rojo build default.project.json -o Place.rbxl
 
-# Success.
+# Success
 echo "Successfully installed project dependencies."
